backend/requirements/testing.txt
factory-boy==2.12.0

# Testing
# ------------------------------------------------------------------------------
pytest==7.4.2
pytest-sugar==0.9.7
pytest-django==4.5.2
pytest-xdist==3.5.0
pytest-mock==3.12.0
pytest-randomly==3.15.0
pytest-freezegun==0.4.2
moto==4.2.14
requests-mock==1.11.0
vcrpy==4.4.0
pytest-env==0.8.2
mock==5.1.0
pytest-socket==0.7.0
pytest-recording==0.13.1
pytest-test-groups==1.0.3


# Coverage Plugins
# ------------------------------------------------------------------------------
pytest-cov==4.1.0
django-coverage-plugin==2.0.4
